+++ packages/chronus/src/config/index.ts
@@ -1,2 +1,3 @@
 export { parseConfig } from "./parse.js";
+export { resolveConfig } from "./resolve.js";
 export type * from "./types.js";
